Important Information for computer use
The commandsGit file contains the most useful commands for using the git
tool.
The commandsVim file contains the most useful commands for using the Vim
tool.
The commandsTmux file contains the most useful commands for using the
Tmux tool.
The usefulCommands.txt file contains information about commandas that
make it easy to perform a task or fix an error.
The installedPrograms.md file contains the list of manually installed
packages or programs.